@@ -1,6 +1,6 @@
 {
   "name": "dash_extendable_graph",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "plotly-dash custom component. Adds Plotly.extendTraces() support to dash_core_components.Graph()",
   "homepage": "https://github.com/bcliang/dash-extendable-graph",
   "main": "build/index.js",
@@ -14,7 +14,7 @@
     "build:py": "dash-generate-components ./src/lib/components dash_extendable_graph",
     "build": "npm run generator && npm run build:js && npm run build:js-dev && npm run build:py",
     "lint": "eslint src",
-    "lint:py": "flake8 --ignore=E501,F401,F841,F811 tests",
+    "lint:py": "flake8 --ignore=E501,W503,F401,F841,F811 tests",
     "format": "prettier --config .prettierrc --write src/**/*.js src/**/**/*.js",
     "test": "run-s -c test:py lint format lint:py",
     "test:py": "pytest tests",
@@ -26,34 +26,35 @@
   },
   "license": "MIT",
   "dependencies": {
-    "ramda": "^0.26.1",
-    "plotly.js": "1.51.1"
+    "plotly.js": "^1.54.5",
+    "ramda": "^0.26.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.4.0",
-    "@babel/preset-env": "^7.5.4",
-    "@babel/preset-react": "^7.0.0",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.0.6",
-    "component-playground": "^3.0.0",
-    "css-loader": "^1.0.1",
-    "eslint": "^6.2.2",
-    "eslint-config-prettier": "^6.1.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-react": "^7.14.3",
+    "@babel/core": "^7.10.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.10.3",
+    "@babel/preset-env": "^7.10.3",
+    "@babel/preset-react": "^7.10.1",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.1.0",
+    "component-playground": "^3.2.1",
+    "css-loader": "^3.6.0",
+    "eslint": "^7.3.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-import": "^2.21.2",
+    "eslint-plugin-react": "^7.20.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.18.2",
+    "prettier": "^2.0.5",
     "prop-types": "^15.7.2",
-    "react": "^16.8.6",
+    "react": "^16.13.0",
     "react-docgen": "^4.1.1",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.13.0",
+    "react-resize-detector": "^4.2.1",
     "style-loader": "^0.23.1",
-    "styled-jsx": "^3.1.1",
-    "webpack": "^4.35.3",
-    "webpack-merge": "^4.2.1",
-    "webpack-nano": "^0.6.2",
-    "webpack-plugin-serve": "^0.11.0"
+    "styled-jsx": "^3.3.0",
+    "webpack": "^4.43.0",
+    "webpack-merge": "^4.2.2",
+    "webpack-nano": "^1.0.0",
+    "webpack-plugin-serve": "^1.0.1"
   },
   "engines": {
     "node": ">=8.11.0",
